LPS-92028 Move and rename
